feat: some fixes on diary fetching with swr and fixes on edit diary
