+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,2 +1,3 @@
 import { handlers } from '@/auth'
+
 export const { GET, POST } = handlers
