+++ src/components/DiaryCard.tsx
@@ -4,6 +4,8 @@ import {
   CalendarIcon,
   EyeIcon,
   EyeSlashIcon,
+  PencilIcon,
+  TrashIcon,
   UserIcon
 } from '@heroicons/react/24/outline'
 import Link from 'next/link'
@@ -13,14 +15,24 @@ interface DiaryCardProps {
   diary: IDiary
   isOwner?: boolean
   featured?: boolean
+  onDelete?: (id: string) => void
 }
 
 const DiaryCard: FC<DiaryCardProps> = ({
   diary,
   isOwner = false,
-  featured = false
+  featured = false,
+  onDelete
 }) => {
   const [contentPreview, setContentPreview] = useState<string>('')
+
+  // Handle delete click
+  const handleDeleteClick = (e: React.MouseEvent) => {
+    e.preventDefault() // Prevent navigation
+    if (onDelete) {
+      onDelete(diary._id)
+    }
+  }
 
   // Log diary data for debugging
   console.log('DiaryCard received:', {
@@ -70,7 +82,7 @@ const DiaryCard: FC<DiaryCardProps> = ({
   }, [diary.content, featured])
 
   return (
-    <Link href={`/diary/${diary._id}`} className="h-full">
+    <div className="h-full">
       <div
         className={`group rounded-xl overflow-hidden bg-white dark:bg-zinc-800 shadow-sm border border-gray-200 dark:border-zinc-700 hover:shadow-md transition duration-300 mb-4 h-full flex flex-col ${
           featured ? 'featured-card' : ''
@@ -81,27 +93,48 @@ const DiaryCard: FC<DiaryCardProps> = ({
             featured ? 'featured-content' : ''
           }`}
         >
-          {/* Header with title and owner indicator */}
-          <div className="mb-1">
-            <h3
-              className={`font-medium text-blue-500 dark:text-blue-400 group-hover:text-blue-600 dark:group-hover:text-blue-500 transition-colors ${
-                featured ? 'text-xl' : 'text-lg'
-              }`}
-            >
-              {diary.title}
-              {isOwner && (
-                <span
-                  className="ml-2 inline-flex items-center"
-                  title="Your diary"
+          {/* Header with title, owner indicator, and action buttons */}
+          <div className="flex justify-between items-start mb-1">
+            <Link href={`/diary/${diary._id}`} className="flex-grow">
+              <h3
+                className={`font-medium text-blue-500 dark:text-blue-400 group-hover:text-blue-600 dark:group-hover:text-blue-500 transition-colors ${
+                  featured ? 'text-xl' : 'text-lg'
+                }`}
+              >
+                {diary.title}
+                {isOwner && (
+                  <span
+                    className="ml-2 inline-flex items-center"
+                    title="Your diary"
+                  >
+                    <UserIcon
+                      className={`text-blue-500 dark:text-blue-400 ${
+                        featured ? 'h-5 w-5' : 'h-4 w-4'
+                      }`}
+                    />
+                  </span>
+                )}
+              </h3>
+            </Link>
+
+            {/* Action buttons for owner */}
+            {isOwner && (
+              <div className="flex items-center gap-2 ml-4">
+                <Link
+                  href={`/diary/${diary._id}/edit`}
+                  className="text-blue-500 hover:text-blue-600 dark:text-blue-400 dark:hover:text-blue-500"
+                  onClick={(e) => e.stopPropagation()}
                 >
-                  <UserIcon
-                    className={`text-blue-500 dark:text-blue-400 ${
-                      featured ? 'h-5 w-5' : 'h-4 w-4'
-                    }`}
-                  />
-                </span>
-              )}
-            </h3>
+                  <PencilIcon className="h-4 w-4" />
+                </Link>
+                <button
+                  onClick={handleDeleteClick}
+                  className="text-red-500 hover:text-red-600 dark:text-red-400 dark:hover:text-red-500"
+                >
+                  <TrashIcon className="h-4 w-4" />
+                </button>
+              </div>
+            )}
           </div>
 
           {/* Meta information row */}
@@ -135,39 +168,41 @@ const DiaryCard: FC<DiaryCardProps> = ({
             </span>
           </div>
 
-          <p
-            className={`text-gray-600 dark:text-gray-300 mb-3 flex-grow ${
-              featured ? 'text-base' : 'text-sm'
-            }`}
-          >
-            {contentPreview}
-          </p>
+          <Link href={`/diary/${diary._id}`} className="flex-grow">
+            <p
+              className={`text-gray-600 dark:text-gray-300 mb-3 ${
+                featured ? 'text-base' : 'text-sm'
+              }`}
+            >
+              {contentPreview}
+            </p>
 
-          {/* Footer with tags and read more */}
-          <div className="flex justify-between items-center mt-3">
-            {/* Tags */}
-            <div className="flex flex-wrap gap-2">
-              {diary.tags && diary.tags.length > 0
-                ? diary.tags.map((tag, index) => (
-                    <span
-                      key={index}
-                      className="text-xs px-2 py-1 rounded-full bg-gray-100 dark:bg-zinc-700 text-gray-700 dark:text-gray-300"
-                    >
-                      {tag}
-                    </span>
-                  ))
-                : null}
-            </div>
+            {/* Footer with tags and read more */}
+            <div className="flex justify-between items-center mt-3">
+              {/* Tags */}
+              <div className="flex flex-wrap gap-2">
+                {diary.tags && diary.tags.length > 0
+                  ? diary.tags.map((tag, index) => (
+                      <span
+                        key={index}
+                        className="text-xs px-2 py-1 rounded-full bg-gray-100 dark:bg-zinc-700 text-gray-700 dark:text-gray-300"
+                      >
+                        {tag}
+                      </span>
+                    ))
+                  : null}
+              </div>
 
-            {/* Read more link */}
-            <div className="text-sm text-blue-500 dark:text-blue-400 font-medium flex items-center opacity-0 group-hover:opacity-100 transition-opacity">
-              Read more
-              <ArrowRightIcon className="ml-1 h-4 w-4" />
+              {/* Read more link */}
+              <div className="text-sm text-blue-500 dark:text-blue-400 font-medium flex items-center opacity-0 group-hover:opacity-100 transition-opacity">
+                Read more
+                <ArrowRightIcon className="ml-1 h-4 w-4" />
+              </div>
             </div>
-          </div>
+          </Link>
         </div>
       </div>
-    </Link>
+    </div>
   )
 }
